+++ source/redux/HttpClient.js
@@ -1,53 +1,43 @@
 import HttpClient from '../HttpClient'
 
-export default function createHttpClient(settings, get_store, options = {})
-{
+export default function createHttpClient(settings, getStore, options = {}) {
 	let on_before_send
 	let catch_to_retry
 	let get_access_token
 
-	// // Add `getState` to `http.request` parameters.
-	// if (settings.http.request)
-	// {
-	// 	on_before_send = (request) =>
-	// 	{
-	// 		// If using Redux, then add `store` as a parameter
-	// 		// for `http_client` customization function
-	// 		settings.http.onBeforeSend(request,
-	// 		{
-	// 			getState: get_store().getState
-	// 		})
-	// 	}
-	// }
+	// Add `getState()` to `http.onRequest()` parameters.
+	if (settings.http.onRequest) {
+		on_before_send = (request, parameters) => {
+			settings.http.onRequest(request, {
+				...parameters,
+				getState: getStore().getState
+			})
+		}
+	}
 
 	// Add `store` and `http` helpers to `http.catch`
-	if (settings.http.catch)
-	{
-		catch_to_retry = (error, retryCount, helpers) =>
-		{
+	if (settings.http.catch) {
+		catch_to_retry = (error, retryCount, helpers) => {
 			return settings.http.catch(error, retryCount, helpers)
 			// {
 			// 	...helpers,
-			// 	store: get_store()
+			// 	store: getStore()
 			// })
 		}
 	}
 
 	// Add `store` helper to `authentication.accessToken`
-	if (settings.authentication.accessToken)
-	{
-		get_access_token = (getCookie, helpers) =>
-		{
+	if (settings.authentication.accessToken) {
+		get_access_token = (getCookie, helpers) => {
 			return settings.authentication.accessToken({
 				...helpers,
 				getCookie,
-				getState: get_store().getState
+				getState: getStore().getState
 			})
 		}
 	}
 
-	return new HttpClient
-	({
+	return new HttpClient({
 		on_before_send,
 		catch_to_retry,
 		get_access_token,